chore: moved data to `data.ts`
